server/requirement.txt
asttokens==2.4.1
attrs==24.2.0
backcall==0.2.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.8.2
cachelib==0.13.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.0.1
fastjsonschema==2.20.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-Session==0.8.0
Flask-SocketIO==5.3.7
Flask-SQLAlchemy==3.1.1
ForgeryPy==0.1
greenlet==3.0.3
h11==0.14.0
idna==3.8
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
msgspec==0.18.6
mysql-connector-python==9.0.0
mysqlclient==2.2.4
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
netifaces==0.10.6
packaging==24.1
pandocfilters==1.5.1
parameterized==0.9.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2024.1
pyzmq==26.2.0
redis==5.0.8
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
setuptools==3.3
simple-websocket==1.0.0
six==1.16.0
socketio==0.2.1
socketIO-client==0.7.2
soupsieve==2.6
SQLAlchemy==2.0.31
stack-data==0.6.3
tinycss2==1.3.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
yarg==0.1.9
